Synchronize results output for cleangoldenami script
